sentiment-platform/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pydantic==2.7.1
python-dotenv==1.0.1
youtube-comment-downloader==0.1.68
requests==2.31.0
beautifulsoup4==4.12.3
vaderSentiment==3.3.2
nltk==3.8.1
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
aiohttp==3.9.5
python-multipart==0.0.9
